requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dateutil==2.9.0.post0
gspread==6.1.2
google-auth==2.29.0
streamlit==1.32.2
requests==2.31.0
